+++ tests/test_import_avro.py
@@ -57,7 +57,7 @@ models:
         required: true
       emailadresses:
         type: array
-        description: Different email adresses of a customer
+        description: Different email addresses of a customer
         items:
            type: string
            format: email
@@ -98,10 +98,10 @@ models:
       orderid:
         type: int
         required: true
-      adresses:
+      addresses:
         type: array
         required: true
-        description: Adresses of a customer
+        description: Addresses of a customer
         items:
           type: object
           fields: